+++ psyke/cart/__init__.py
@@ -1,4 +1,4 @@
-from psyke.cart.cart_predictor import CartPredictor, LeafConstraints, LeafSequence
+from psyke.cart.predictor import CartPredictor, LeafConstraints, LeafSequence
 from psyke import Extractor, DiscreteFeature
 from psyke.schema import GreaterThan
 from psyke.utils.logic_utils import create_variable_list, create_head, create_term
